backend/requirements.txt
# NewsGap Backend Dependencies

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
aiosqlite==0.20.0

# HTTP Client & Crawling
httpx==0.27.2
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0
lxml_html_clean==0.3.1
readability-lxml==0.8.1
python-dateutil==2.9.0

# LLM
openai==1.54.4
google-generativeai==0.8.3
tiktoken==0.8.0

# Data Validation
pydantic==2.9.2
pydantic-settings==2.6.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
